backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.7.1
pymongo==4.11.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
google-genai==1.41.0
redis==5.2.0
